requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
anthropic>=0.40.0
python-dotenv==1.0.1
